+++ src/__tests__/integration/json/arrayContainsMultiple.spec.ts
@@ -1,7 +1,7 @@
 import './setup';
 import { prisma } from './setup';
 import { nanoid } from 'nanoid';
-import { buildQuery } from '../../query-builder';
+import { buildQuery } from '../../../query-builder';
 
 describe('Array Contains Multiple Elements', () => {
   let ids: Record<string, string> = {};
