+++ src/__tests__/integration/json/setup.ts
@@ -1,0 +1,4 @@
+import '../setup';
+import { prisma } from '../setup';
+
+export { prisma };
